Expose aggregations as carto.operation
