+++ src/api/v4/dataview/index.js
@@ -1,7 +1,5 @@
 var Formula = require('./formula');
-var AGGREGATIONS = require('./constants').AGGREGATIONS;
 
 module.exports = {
-  Formula: Formula,
-  AGGREGATIONS: AGGREGATIONS
+  Formula: Formula
 };
