+++ src/api/v4/index.js
@@ -1,12 +1,12 @@
 /**
  *  @api
  *  @namespace carto
- *  
+ *
  *  @description
  *  # Carto.js
  *  All the library features are exposed through the `carto` namespace.
- *  
- * 
+ *
+ *
  * - **Client** : The api client.
  * - **Events** : The events exposed.
  * - **source** : Source description
@@ -18,6 +18,7 @@ var style = require('./style');
 var layer = require('./layer');
 var Events = require('./events');
 var dataview = require('./dataview');
+var OPERATION = require('./constants').OPERATION;
 
 module.exports = {
   Client: Client,
@@ -25,5 +26,6 @@ module.exports = {
   source: source,
   style: style,
   layer: layer,
-  dataview: dataview
+  dataview: dataview,
+  operation: OPERATION
 };
